Moved Kurucz partfunc test
